pyproject.toml
[project]
name = "kafka-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "confluent-kafka[avro]>=2.10.0",
    "dotenv>=0.9.9",
    "faker>=37.1.0",
    "faker-vehicle>=0.2.0",
    "kafka-python==2.1.5",
    "psycopg2-binary>=2.9.10",
    "requests==2.32.3",
]
